Control core sequence's transaction count from top in ICache tests
This will have no effect for now: we just move the "pick a number in
the range 800..1000" logic to the virtual sequence.

The reason to do this is for tests that combine sequences: we want to
be able to shorten each component sequence so that the combined test
isn't way longer than the original ones were.